Updating to latest FlutterFlow output.

+++ lib/gestion_cartes/the_main_mes_cartes_dashboard_view_cards_transactions/the_main_mes_cartes_dashboard_view_cards_transactions_widget.dart
@@ -1953,28 +1953,31 @@ class _TheMainMesCartesDashboardViewCardsTransactionsWidgetState
                                                             },
                                                           ) ??
                                                           false;
-                                                  await listViewTransactionsRecord
-                                                      .reference
-                                                      .delete();
-                                                  ScaffoldMessenger.of(context)
-                                                      .showSnackBar(
-                                                    SnackBar(
-                                                      content: Text(
-                                                        'Transaction Deleted Successfully',
-                                                        style: TextStyle(
-                                                          color: FlutterFlowTheme
-                                                                  .of(context)
-                                                              .primaryText,
+                                                  if (confirmDialogResponse) {
+                                                    await listViewTransactionsRecord
+                                                        .reference
+                                                        .delete();
+                                                    ScaffoldMessenger.of(
+                                                            context)
+                                                        .showSnackBar(
+                                                      SnackBar(
+                                                        content: Text(
+                                                          'Transaction Deleted Successfully',
+                                                          style: TextStyle(
+                                                            color: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .primaryText,
+                                                          ),
                                                         ),
+                                                        duration: Duration(
+                                                            milliseconds: 4000),
+                                                        backgroundColor:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .secondary,
                                                       ),
-                                                      duration: Duration(
-                                                          milliseconds: 4000),
-                                                      backgroundColor:
-                                                          FlutterFlowTheme.of(
-                                                                  context)
-                                                              .secondary,
-                                                    ),
-                                                  );
+                                                    );
+                                                  }
                                                 },
                                                 child: Container(
                                                   width: double.infinity,
